requirements.txt
black==24.10.0
certifi==2024.12.14
click==8.1.7
distlib==0.3.9
filelock==3.16.1
flake8==7.1.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pipenv==2024.4.0
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.9.4
tomli==2.2.1
typing_extensions==4.12.2
utility==1.0
virtualenv==20.28.0
pydantic~=2.10.4